Code refactor, merging all User type in one Object
