--- src/main/java/fr/diginamic/app/dto/AdminDto.java
@@ -1,4 +1,0 @@
-package fr.diginamic.app.dto;
-
-public class AdminDto extends EmployeDto{
-}
