--- src/main/java/fr/diginamic/app/dto/EmployeDto.java
@@ -1,6 +1,0 @@
-package fr.diginamic.app.dto;
-
-public class EmployeDto extends UserDto {
-    private float daysoffBalance;
-    private float emplRttBalance;
-}
