--- src/main/java/fr/diginamic/app/dto/ManagerDto.java
@@ -1,4 +1,0 @@
-package fr.diginamic.app.dto;
-
-public class ManagerDto extends EmployeDto {
-}
