--- src/main/java/fr/diginamic/app/repository/EmployeRepository.java
@@ -1,12 +1,0 @@
-package fr.diginamic.app.repository;
-
-import fr.diginamic.app.model.Employe;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface EmployeRepository extends JpaRepository<Employe, Long> {
-    Optional<Employe> findByEmail(String email);
-}
