+++ src/main/java/fr/diginamic/app/repository/EmployeeRepository.java
@@ -1,0 +1,12 @@
+package fr.diginamic.app.repository;
+
+import fr.diginamic.app.model.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+    Optional<Employee> findByEmail(String email);
+}
