--- src/main/java/fr/diginamic/app/repository/ManagerRepository.java
@@ -1,12 +1,0 @@
-package fr.diginamic.app.repository;
-
-import fr.diginamic.app.model.Manager;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface ManagerRepository extends JpaRepository<Manager, Long> {
-    Optional<Manager> findByEmail(String email);
-}
