+++ src/main/java/fr/diginamic/app/service/impl/EmployeeServiceImpl.java
@@ -1,0 +1,37 @@
+package fr.diginamic.app.service.impl;
+
+import fr.diginamic.app.model.Employee;
+import fr.diginamic.app.repository.EmployeeRepository;
+import fr.diginamic.app.service.EmployeeService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class EmployeeServiceImpl implements EmployeeService {
+
+    @Autowired
+    private EmployeeRepository employeeRepository;
+
+    @Override
+    public Employee save(Employee user) {
+        return employeeRepository.save(user);
+    }
+
+    @Override
+    public List<Employee> findAll() {
+        return employeeRepository.findAll();
+    }
+
+    @Override
+    public Optional<Employee> findById(Long id) {
+        return employeeRepository.findById(id);
+    }
+
+    @Override
+    public void delete(Long id) {
+        employeeRepository.deleteById(id);
+    }
+}
