--- src/main/java/fr/diginamic/app/service/impl/UserServiceImpl.java
@@ -1,37 +1,0 @@
-package fr.diginamic.app.service.impl;
-
-import fr.diginamic.app.model.User;
-import fr.diginamic.app.repository.UserRepository;
-import fr.diginamic.app.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class UserServiceImpl implements UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public User save(User user) {
-        return userRepository.save(user);
-    }
-
-    @Override
-    public List<User> findAll() {
-        return userRepository.findAll();
-    }
-
-    @Override
-    public Optional<User> findById(Long id) {
-        return userRepository.findById(id);
-    }
-
-    @Override
-    public void delete(Long id) {
-        userRepository.deleteById(id);
-    }
-}
